REFACTORING:: modal 폴더구조 벼경
