+++ src/layout/LoginModal/modal.tsx
@@ -1,7 +1,7 @@
 import React, { useRef, useState } from "react";
 import styled from "styled-components";
-import { useLoginModal } from "../hooks/useLoginMdal";
-import { useOutSideClick } from "../hooks/useOutsideClick";
+import { useLoginModal } from "../../hooks/useLoginMdal";
+import { useOutSideClick } from "../../hooks/useOutsideClick";
 import LoginModalPage from "./login";
 import JoinInfinity from "./join";
